drivers/gpu/arm/coresight/sources/itm/coresight_mali_source_itm_core.c
drivers/gpu/arm/midgard/debug/backend/mali_kbase_debug_ktrace_jm.c
drivers/gpu/arm/midgard/device/backend/mali_kbase_device_hw_jm.c
drivers/gpu/arm/midgard/hw_access/regmap/mali_kbase_regmap_jm_enums.h
drivers/gpu/arm/midgard/hw_access/regmap/mali_kbase_regmap_jm_macros.h
drivers/gpu/arm/midgard/hwcnt/backend/mali_kbase_hwcnt_backend_jm.c
drivers/gpu/arm/midgard/hwcnt/backend/mali_kbase_hwcnt_backend_jm_watchdog.c
drivers/gpu/arm/midgard/mmu/backend/mali_kbase_mmu_faults_decoder_luts_jm.h
drivers/gpu/arm/valhall-inc/uapi/gpu/arm/midgard/jm/mali_base_jm_kernel.h
drivers/gpu/arm/valhall-inc/uapi/gpu/arm/midgard/jm/mali_kbase_jm_ioctl.h

+++ drivers/bluetooth/btmtkuart.c
@@ -8,7 +8,7 @@
  *
  */
 
-#include <asm/unaligned.h>
+#include <linux/unaligned.h>
 #include <linux/atomic.h>
 #include <linux/clk.h>
 #include <linux/firmware.h>
@@ -384,8 +384,8 @@ static void btmtkuart_recv(struct hci_dev *hdev, const u8 *data, size_t count)
 	}
 }
 
-static int btmtkuart_receive_buf(struct serdev_device *serdev, const u8 *data,
-				 size_t count)
+static size_t btmtkuart_receive_buf(struct serdev_device *serdev,
+				    const u8 *data, size_t count)
 {
 	struct btmtkuart_dev *bdev = serdev_device_get_drvdata(serdev);
